# internal/openapi/swagger.yaml
basePath: /api
definitions:
  api.APIEnvelope:
    properties:
      data:
        type: object
      message:
        example: ok
        type: string
      success:
        example: true
        type: boolean
    type: object
  api.APIErrorEnvelope:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: true
        type: boolean
      message:
        example: invalid request
        type: string
    type: object
...
  api.ManagedPeerStatsDTO:
    properties:
      endpoint:
        example: 5.6.7.8:54321
        type: string
      lastHandshake:
        example: "2024-01-15T10:30:00Z"
        type: string
      online:
        example: true
        type: boolean
      publicKey:
        example: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF=
        type: string
      rxBytes:
        example: 2097152
        type: integer
      txBytes:
        example: 1048576
        type: integer
    type: object
...
  api.ManagedServerStatsDTO:
    properties:
      peers:
        items:
          $ref: '#/definitions/api.ManagedPeerStatsDTO'
        type: array
      status:
        example: up
        type: string
    type: object
...
  api.ServersAllData:
    properties:
      managedStats:
        $ref: '#/definitions/api.ManagedServerStatsDTO'
      servers:
        items:
          $ref: '#/definitions/api.WireguardServerDTO'
        type: array
      wanIP:
        example: 203.0.113.42
        type: string
    type: object
  api.ServersAllResponse:
    properties:
      data:
        $ref: '#/definitions/api.ServersAllData'
      success:
        example: true
        type: boolean
    type: object
...
  api.WANIPData:
    properties:
      ip:
        example: 203.0.113.42
        type: string
    type: object
  api.WANIPResponse:
    properties:
      data:
        $ref: '#/definitions/api.WANIPData'
      success:
        example: true
        type: boolean
    type: object
...
  api.WireguardServerDTO:
    properties:
      address:
        example: 10.0.1.1
        type: string
      connected:
        example: true
        type: boolean
      description:
        example: Wireguard VPN Server
        type: string
      id:
        example: Wireguard0
        type: string
      interfaceName:
        example: Wireguard0
        type: string
      listenPort:
        example: 51820
        type: integer
      mask:
        example: 255.255.255.0
        type: string
      mtu:
        example: 1420
        type: integer
      peers:
        items:
          $ref: '#/definitions/api.WireguardServerPeerDTO'
        type: array
      publicKey:
        example: EEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEE=
        type: string
      status:
        example: up
        type: string
    type: object
  api.WireguardServerPeerDTO:
    properties:
      allowedIPs:
        example:
        - 10.0.1.2/32
        items:
          type: string
        type: array
      description:
        example: Phone
        type: string
      enabled:
        example: true
        type: boolean
      endpoint:
        example: 1.2.3.4:12345
        type: string
      lastHandshake:
        example: "2024-01-15T10:30:00Z"
        type: string
      online:
        example: true
        type: boolean
      publicKey:
        example: DDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD=
        type: string
      rxBytes:
        example: 1048576
        type: integer
      txBytes:
        example: 524288
        type: integer
    type: object
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /servers/all:
    get:
      description: 'Returns the composite servers snapshot: WireGuard servers list,
        managed server, stats and WAN IP.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ServersAllResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Get all servers snapshot
      tags:
      - servers
  /servers/mark:
    delete:
      description: POST marks the named WG interface as a server (visible under Servers,
        hidden from Tunnels). DELETE unmarks (returns it to the Tunnels list). Both
        return the fresh ServersSnapshot.
      parameters:
      - description: Interface name (e.g. Wireguard0)
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ServersAllResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Mark/unmark interface as server
      tags:
      - servers
    post:
      description: POST marks the named WG interface as a server (visible under Servers,
        hidden from Tunnels). DELETE unmarks (returns it to the Tunnels list). Both
        return the fresh ServersSnapshot.
      parameters:
      - description: Interface name (e.g. Wireguard0)
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ServersAllResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Mark/unmark interface as server
      tags:
      - servers
  /servers/marked:
    get:
      description: Returns the list of interface IDs that have been marked as servers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Get marked server interfaces
      tags:
      - servers
  /servers/wan-ip:
    get:
      description: Returns the router's external WAN IP address.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.WANIPResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Get WAN IP
      tags:
      - servers
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"